+++ exporter/exporterhelper/metrics.go
@@ -79,10 +79,21 @@ type metricsExporter struct {
 	consumer.Metrics
 }
 
-// NewMetricsExporter creates an MetricsExporter that records observability metrics and wraps every request with a Span.
+// Deprecated: [v0.58.0] use NewMetricsExporterWithContext.
 func NewMetricsExporter(
 	cfg config.Exporter,
 	set component.ExporterCreateSettings,
+	pusher consumer.ConsumeMetricsFunc,
+	options ...Option,
+) (component.MetricsExporter, error) {
+	return NewMetricsExporterWithContext(context.Background(), set, cfg, pusher, options...)
+}
+
+// NewMetricsExporterWithContext creates an MetricsExporter that records observability metrics and wraps every request with a Span.
+func NewMetricsExporterWithContext(
+	_ context.Context,
+	set component.ExporterCreateSettings,
+	cfg config.Exporter,
 	pusher consumer.ConsumeMetricsFunc,
 	options ...Option,
 ) (component.MetricsExporter, error) {